Move type checker tests to aicsimage tests
